+++ src/main/java/wooteco/subway/domain/line/Line.java
@@ -1,6 +1,8 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.line;
 
 import java.util.List;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.section.Sections;
 
 public class Line {
     private Long id;
