+++ src/main/java/wooteco/subway/domain/section/Distance.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 public class Distance {
     public static final int MINIMUM_DISTANCE = 1;
