+++ src/main/java/wooteco/subway/domain/section/Section.java
@@ -1,8 +1,9 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
+import wooteco.subway.domain.station.Station;
 
 public class Section {
     private final Long id;
