+++ src/main/java/wooteco/subway/domain/station/Station.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.station;
 
 import java.util.Objects;
 import wooteco.subway.dto.StationRequest;
