+++ src/main/java/wooteco/subway/dto/StationResponse.java
@@ -1,6 +1,6 @@
 package wooteco.subway.dto;
 
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.station.Station;
 
 public class StationResponse {
     private Long id;
