+++ src/main/java/wooteco/subway/service/PathService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Path;
-import wooteco.subway.domain.Section;
-import wooteco.subway.domain.Station;
+import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.PathResponse;
 import wooteco.subway.exception.NotFoundStationException;
